+++ CHANGELOG.md
@@ -29,6 +29,7 @@ and this project adheres to
 - all: changes panic! to panic_with_error ([#269])
 - Audit Report from Veridise ([#291])
 - all: added entrypoint to update the contracts ([#293])
+- Pool and Pool Stable: adds query for totally issued tokens ([#298])
 
 [#267]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/267
 [#269]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/269
@@ -39,6 +40,7 @@ and this project adheres to
 [#289]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/289
 [#291]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/291
 [#293]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/293
+[#298]: https://github.com/Phoenix-Protocol-Group/phoenix-contracts/pull/298
 
 ## Bug fixes